Runtime and Engine Refactor for Testing
